Adding osu and wybin pages
