+++ extension/src/ui/options/content/Content.tsx
@@ -1,4 +1,5 @@
 import { useHashContext } from "../context/hashContext";
+import { TabContextProvider } from "../context/tabContext";
 import Settings from "./pages/Settings";
 import { Match, Switch } from "solid-js";
 
@@ -6,11 +7,12 @@ export default function Content() {
     const { hash } = useHashContext();
 
     return (
-        <div class="p-4 sm:ml-64 max-w-3xl">
-            <Switch>
-                <Match when={"" === hash()}>{<Settings />}</Match>
-                <Match when={"#settings" === hash()}>{<Settings />}</Match>
-            </Switch>
-        </div>
+        <TabContextProvider>
+            <div class="p-4 sm:ml-64 max-w-3xl">
+                <Switch>
+                    <Match when={["#settings", ""].includes(hash())}>{<Settings />}</Match>
+                </Switch>
+            </div>
+        </TabContextProvider>
     );
 }
