+++ extension/src/ui/options/content/pages/subpages/WybinSettings.tsx
@@ -1,0 +1,5 @@
+import SettingsLayout from "../SettingsLayout";
+
+export default function WybinSettings() {
+    return <SettingsLayout websiteName="wybin" link="https://wybin.xyz/"></SettingsLayout>;
+}
